все функции перенес в один файл helpers.php, сделал strip_tags на вывод, мелкие правки
